+++ AnimalsNeedWater/Types/ModData.cs
@@ -9,5 +9,8 @@ namespace AnimalsNeedWater.Types
         public static List<string> BuildingsWithWateredTrough { get; set; } = new List<string>();
         public static List<FarmAnimal> FullAnimals { get; set; } = new List<FarmAnimal>();
         public static int LoveEmote { get; } = 20;
+        
+        public static readonly string WaterBowlItemId = $"{ModEntry.ModHelper.ModContent.ModID}_Water_Bowl";
+        public static readonly string WaterBowlItemModDataIsFullField = $"{ModEntry.ModHelper.ModContent.ModID}_Water_Bowl.IsFull";
     }
 }